refactor: ChessGameController 로 변경
